issue persist

+++ src/modules/foodspot/services/foodspot.service.ts
@@ -1,19 +1,31 @@
-import { FoodspotRepository, FoodspotInput,FoosspotUpdateInput} from "../repositories/foodspot.repository.js";
+import {
+  FoodspotInput,
+  FoodspotUpdateInput,
+} from "../../../types/foodspot.types";
+import { FoodspotRepository } from "../repositories/foodspot.repository";
 
 export const FoodspotService = {
-  createFoodspot: async (data:FoodspotInput) => {
-    return await FoodspotRepository.create(data);
+  createFoodspot: async (input: FoodspotInput, ownerId: string) => {
+    return await FoodspotRepository.create(input, ownerId);
   },
   getAllFoodspot: async () => {
     return await FoodspotRepository.getAll();
   },
-  getFoodspotByid: async (id:string) => {
+  getFoodspotByid: async (id: string) => {
     return await FoodspotRepository.getById(id);
   },
-  updateFoodspot:async(id:string,data:FoosspotUpdateInput)=>{
-    return await FoodspotRepository.update(id,data)
+  updateFoodspot: async (
+    id: string,
+    data: FoodspotUpdateInput,
+    ownerId: string
+  ) => {
+    const foodspot = await FoodspotRepository.findByIdAndOwner(id, ownerId);
+    if (!foodspot) throw new Error("Unauthorized or not found");
+    return await FoodspotRepository.update(id, data);
   },
-  deleteFoodspot:async(id:string)=>{
-    return await FoodspotRepository.delete(id)
-  }
+  deleteFoodspot: async (id: string, ownerId: string) => {
+    const foodspot = await FoodspotRepository.findByIdAndOwner(id, ownerId);
+    if (!foodspot) throw new Error("Unauthorized or not found");
+    return await FoodspotRepository.delete(id);
+  },
 };
